+++ OMG.PSUtilities.AI/OMG.PSUtilities.AI.psm1
@@ -18,6 +18,7 @@ Get-ChildItem -Path "$PSScriptRoot\Public\*.ps1" -Recurse | Where-Object{$_.name
 
 # Export public functions
 $PublicFunctions = @(
+    'Convert-PSUContext',
     'Get-PSUAiPoweredGitChangeSummary',
     'Invoke-PSUAiPrompt',
     'Invoke-PSUGitCommit',
@@ -36,7 +37,9 @@ $AliasesToExport = @(
     'aigitcommit',
     'askai',
     'askazureopenai',
-    'askgemini'
+    'askgemini',
+    'askperplexity',
+    'rephrase'
 )
 
 Export-ModuleMember -Function $PublicFunctions -Alias $AliasesToExport